update workout api calls and logic for updating and adding new workout, still need to do exercises logic and there is lag somewhere

+++ src/components/Auth/Workouts/parts/Workouts.jsx
@@ -2,32 +2,36 @@ import React from 'react';
 import { withTheme, List } from 'react-native-paper';
 import { View, Animated, Easing } from 'react-native';
 import { FontAwesome5 } from '@expo/vector-icons';
+import { StateContext } from '../../../../controllers/state';
 import CardWithButton from '../../../../template/CardWithButton';
 import ScrollList from '../../../../template/ScrollList';
 import WorkoutItem from './WorkoutItem';
+import WorkoutAPI from '../../../../controllers/WorkoutApi';
 
 // const Data = [];
 const Workouts = ({
   theme,
-  userUid,
   navigation,
   setMessage,
   setNotifyTitle,
   setShowNotify,
   isOk,
   setIsOk,
-  data,
-  setData,
-  setSelectedWorkout,
-  getWorkouts,
 }) => {
+  const {
+    user,
+    selectedWorkout: {selectedWorkout,  setSelectedWorkout },
+    workouts: { workouts, setWorkouts },
+  } = React.useContext(StateContext);
+
   const [isLoading, setIsLoading] = React.useState(false);
-  const [isWorkoutCalled, setIsWorkoutCalled] = React.useState(false);
   const [isDisable, setIsDisable] = React.useState(true);
   const [selected, setSelected] = React.useState(null);
+  const [isDelete, setIsDelete] = React.useState(false);
+  const [isEdit, setIsEdit] = React.useState(false);
   const [modalOnOkSelectedId, setModalOnOkSelected] = React.useState('');
   const springAnim = React.useRef(new Animated.Value(1)).current;
-
+  const isMounted = React.useRef(true);
   const springOut = (callback) => {
     Animated.timing(springAnim, {
       toValue: 0,
@@ -42,22 +46,42 @@ const Workouts = ({
     outputRange: [500, 0],
   });
 
-  const deleteWorkout = (id) => {
-    springOut(() => {
-      setIsOk(false);
-      setModalOnOkSelected(null);
-      springAnim.setValue(1);
-      setData(data.filter((d) => d.title !== id));
-      setSelected(null);
-    });
-    // TODO API call
+  const setUpdatedWorkout = (id) => {
+    if (id) {
+      setSelectedWorkout(workouts.find((a) => a.title === id));
+    } else {
+      setSelectedWorkout();
+    }
   };
 
-  const handleOnSubmit = () => {
+  const deleteWorkout = React.useCallback(
+    (id) => {
+      if (id && user && user.uid && isMounted.current) {
+        WorkoutAPI.deleteWorkout(user.uid, id);
+      }
+      springOut(() => {
+        if (isMounted.current) {
+          setIsOk(false);
+          setModalOnOkSelected(null);
+          setSelected(null);
+        }
+        springAnim.setValue(1);
+        setWorkouts(() => [...workouts].filter((d) => d.id !== id));
+      });
+    },
+    [user, workouts]
+  );
+
+  const editWorkout = React.useCallback(() => {
+    setUpdatedWorkout(selected);
+    navigation.navigate('Create');
+  }, [selected, user, workouts]);
+
+  const handleOnSubmit = React.useCallback(() => {
     setIsLoading(!isDisable);
-    setSelectedWorkout(data[selected]);
+    setUpdatedWorkout(selected);
     navigation.navigate('Exercises');
-  };
+  }, [selected, workouts, isDisable]);
 
   const onPress = (id) => {
     setIsDisable(false);
@@ -66,14 +90,22 @@ const Workouts = ({
 
   const handleEdit = (_, id) => {
     setSelected(id);
+    setIsDisable(true);
+    setNotifyTitle('You wanna change this?');
+    setMessage(`You're about to edit ${workouts.filter((d) => d.id === id)[0].id}, ok?`);
+    setIsOk(false);
+    setIsEdit(true);
+    setModalOnOkSelected(id);
+    setShowNotify(true);
   };
 
   const handleTrash = (_, id) => {
     setIsDisable(true);
     setSelected(id);
     setNotifyTitle('Woah, you sure...');
-    setMessage(`Do you really want to delete ${data.filter((d) => d.title === id)[0].title}?`);
+    setMessage(`Do you really want to delete ${workouts.filter((d) => d.id === id)[0].id}?`);
     setIsOk(false);
+    setIsDelete(true);
     setModalOnOkSelected(id);
     setShowNotify(true);
   };
@@ -82,27 +114,37 @@ const Workouts = ({
     navigation.navigate('Create');
   };
 
-  //  React.useEffect(() => {
-  //    if (springAnim === 1) {
-  //    }
-  //  }, [springAnim]);
+  React.useEffect(() => {
+    isMounted.current = true;
+    return () => {
+      isMounted.current = false;
+    };
+  }, []);
 
   React.useEffect(() => {
-    if (isOk && modalOnOkSelectedId) {
-      deleteWorkout(modalOnOkSelectedId);
+    if (isMounted.current) {
+      if (isOk && isDelete && modalOnOkSelectedId) {
+        deleteWorkout(modalOnOkSelectedId);
+      } else if (isOk && isEdit && modalOnOkSelectedId) {
+        editWorkout(modalOnOkSelectedId);
+      }
+      setIsDelete(false);
+      setIsEdit(false);
+      setModalOnOkSelected(null);
     }
-    setModalOnOkSelected(null);
   }, [isOk]);
 
   const Item = ({ item }) => (
     <Animated.View
       style={
-        item.id === selected ? { opacity: springAnim, transform: [{ translateX: panX }] } : null
+        item && item.id === selected
+          ? { opacity: springAnim, transform: [{ translateX: panX }] }
+          : null
       }
     >
       <WorkoutItem
-        onPress={() => onPress(item.title)}
-        isSelected={item.title === selected}
+        onPress={() => onPress(item.id)}
+        isSelected={item && item.id === selected}
         text={item}
         handleTrash={handleTrash}
         handleEdit={handleEdit}
@@ -163,21 +205,21 @@ const Workouts = ({
   return (
     <CardWithButton
       buttonText="Select"
-      showButton={data.length > 0}
+      showButton={workouts && workouts.length > 0}
       theme={theme}
       buttonDisabled={isDisable}
       onPress={handleOnSubmit}
       isLoading={isLoading}
       flex={1}
       style={{
-        flex: data.length > 0 ? 2 : null,
+        flex: workouts && workouts.length > 0 ? 2 : null,
         marginBottom: 50,
       }}
     >
       <ScrollList
-        data={data}
+        data={workouts}
         renderItem={Item}
-        keyExtractor={(item) => item.title}
+        keyExtractor={(item) => item && item.id}
         extraData={selected}
         theme={theme}
         ItemSeparatorComponent={ItemSeparator}
